move FileListener to localstack.utils.tail
